% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function3.R
\name{barplot_funct}
\alias{barplot_funct}
\title{Bar Plot Function}
\usage{
barplot_funct(matrix)
}
\arguments{
\item{matrix}{the mean Matrix}
}
\description{
Plots the mean genotype smaples on a barplot
}
\examples{
barplot_funct()
}
\keyword{barplot}
\keyword{plot,}
