% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function2.R
\name{ColMean_funct}
\alias{ColMean_funct}
\title{Mean Cell Type}
\usage{
ColMean_funct(df)
}
\arguments{
\item{df}{the dataframe}
}
\description{
Finds the mean of each cell type per genotype sample
}
\examples{
ColMean_funct()
}
\keyword{colmeans}
\keyword{mean,}
