% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function1.R
\name{Split_Gen_Funct}
\alias{Split_Gen_Funct}
\title{Separate by Genotype}
\usage{
Split_Gen_Funct(df, vect)
}
\arguments{
\item{vect}{by genotype}

\item{input}{the dataframe}
}
\description{
Splits dataframe by genotype
}
\examples{
Split_Gen_Funct()
}
\keyword{genotype,}
\keyword{split,}
\keyword{vect}
